general: Fix logging of pointer-like values

+++ drivers/gfx/bochs/src/main.cpp
@@ -217,7 +217,7 @@ GfxDevice::createDumb(uint32_t width, uint32_t height, uint32_t bpp) {
 	auto offset = _vramAllocator.allocate(alignment + size);
 	auto displacement = alignment - (offset % alignment);
 	if(logBuffers)
-		std::cout << "gfx-bochs: Allocating buffer of size " << (void *)size
+		std::cout << "gfx-bochs: Allocating buffer of size " << (void *)(size_t)size
 				<< " at " << (void *)offset
 				<< ", displacement is: " << (void *)displacement << std::endl;
 	auto buffer = std::make_shared<BufferObject>(this, alignment, size,
